+++ src/main/groovy/com/avioconsulting/mule/testing/transformers/xml/XMLGroovyParserTransformer.groovy
@@ -1,12 +1,14 @@
 package com.avioconsulting.mule.testing.transformers.xml
 
 import com.avioconsulting.mule.testing.payloadvalidators.IPayloadValidator
+import com.avioconsulting.mule.testing.transformers.ClosureMuleMessageHandler
 import groovy.xml.XmlUtil
 import org.mule.api.MuleContext
 import org.mule.api.MuleMessage
 import org.mule.modules.interceptor.processors.MuleMessageTransformer
 
-class XMLGroovyParserTransformer extends XMLTransformer implements MuleMessageTransformer {
+class XMLGroovyParserTransformer extends XMLTransformer implements MuleMessageTransformer,
+        ClosureMuleMessageHandler{
     private final Closure closure
 
     XMLGroovyParserTransformer(Closure closure,
@@ -20,7 +22,8 @@ class XMLGroovyParserTransformer extends XMLTransformer implements MuleMessageTr
         validateContentType(incomingMessage)
         def xmlString = incomingMessage.payloadAsString
         def node = new XmlParser().parseText(xmlString) as Node
-        def reply = closure(node)
+        def forMuleMsg = withMuleMessage(closure, incomingMessage)
+        def reply = forMuleMsg(node)
 
         String outputXmlString
         if (reply instanceof File) {